# Scope: remove fmt.Sprintf() string concatenation and improve tag merge