--- a/tags.go
+++ b/tags.go
@@ -24,15 +24,19 @@ func (t tagSet) Len() int           { return len(t) }
 func (t tagSet) Swap(i, j int)      { t[i], t[j] = t[j], t[i] }
 func (t tagSet) Less(i, j int) bool { return t[i].dimension < t[j].dimension }
 
-func serializeTags(tags map[string]string) string {
+func serializeTags(name string, tags map[string]string) string {
 	const prefix = ".__"
 	const sep = "="
 
 	if len(tags) == 0 {
-		return ""
+		return name
 	}
-	pairs := make([]tagPair, 0, len(tags))
+
+	// n stores the length of the serialized name + tags
 	n := (len(prefix) + len(sep)) * len(tags)
+	n += len(name)
+
+	pairs := make([]tagPair, 0, len(tags))
 	for k, v := range tags {
 		n += len(k) + len(v)
 		pairs = append(pairs, tagPair{
@@ -44,6 +48,7 @@ func serializeTags(tags map[string]string) string {
 
 	// CEV: this is same as strings.Builder, but works with go1.9 and earlier
 	b := make([]byte, 0, n)
+	b = append(b, name...)
 	for _, tag := range pairs {
 		b = append(b, prefix...)
 		b = append(b, tag.dimension...)