--- a/tags_test.go
+++ b/tags_test.go
@@ -5,38 +5,47 @@ import (
 )
 
 func TestSerializeTags(t *testing.T) {
+	const name = "prefix"
+	const expected = name + ".__q=r.__zzz=hello"
 	tags := map[string]string{"zzz": "hello", "q": "r"}
-	serialized := serializeTags(tags)
-	if serialized != ".__q=r.__zzz=hello" {
+	serialized := serializeTags(name, tags)
+	if serialized != expected {
 		t.Errorf("Serialized output (%s) didn't match expected output", serialized)
 	}
 }
 
 func TestSerializeWithPerInstanceFlag(t *testing.T) {
+	const name = "prefix"
+	const expected = name + ".___f=i.__foo=bar"
 	tags := map[string]string{"foo": "bar", "_f": "i"}
-	serialized := serializeTags(tags)
-	if serialized != ".___f=i.__foo=bar" {
+	serialized := serializeTags(name, tags)
+	if serialized != expected {
 		t.Errorf("Serialized output (%s) didn't match expected output", serialized)
 	}
 }
 
 func TestSerializeIllegalTags(t *testing.T) {
+	const name = "prefix"
+	const expected = name + ".__foo=b_a_r.__q=p"
 	tags := map[string]string{"foo": "b|a:r", "q": "p"}
-	serialized := serializeTags(tags)
-	if serialized != ".__foo=b_a_r.__q=p" {
+	serialized := serializeTags(name, tags)
+	if serialized != expected {
 		t.Errorf("Serialized output (%s) didn't match expected output", serialized)
 	}
 }
 
 func TestSerializeTagValuePeriod(t *testing.T) {
+	const name = "prefix"
+	const expected = name + ".__foo=blah_blah.__q=p"
 	tags := map[string]string{"foo": "blah.blah", "q": "p"}
-	serialized := serializeTags(tags)
-	if serialized != ".__foo=blah_blah.__q=p" {
+	serialized := serializeTags(name, tags)
+	if serialized != expected {
 		t.Errorf("Serialized output (%s) didn't match expected output", serialized)
 	}
 }
 
 func BenchmarkSerializeTags(b *testing.B) {
+	const name = "prefix"
 	tags := map[string]string{
 		"tag1": "val1",
 		"tag2": "val2",
@@ -46,6 +55,6 @@ func BenchmarkSerializeTags(b *testing.B) {
 	}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		serializeTags(tags)
+		serializeTags(name, tags)
 	}
 }