This builds on the work of #50.

Improve the performance of Scope* methods by not using fmt.Sprintf() to concatenate strings and re-using the parent's tags when the child scope has none (tags are never modified - they are only used to generate the stat key name).

Benchmark Results:

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkStore_ScopeWithTags-8     1417          1268          -10.52%
BenchmarkStore_ScopeNoTags-8       1100          527           -52.09%

benchmark                          old allocs     new allocs     delta
BenchmarkStore_ScopeWithTags-8     6              4              -33.33%
BenchmarkStore_ScopeNoTags-8       8              4              -50.00%

benchmark                          old bytes     new bytes     delta
BenchmarkStore_ScopeWithTags-8     576           544           -5.56%
BenchmarkStore_ScopeNoTags-8       672           304           -54.76%
```